Added the QueryResult class, and visualization methods for it.
This commit introduces an execute() method in the Query class. This
returns a QueryResult object containing the results of the query. This
class provides users with visualization methods. Currently, we have a
high level method 'plot' that takes in the visualization kind plus
linechart and heatmap to call the methods directly.